+++ src/main/org/excelsi/nausicaa/ca/AbstractFloatPlane.java
@@ -12,13 +12,13 @@ public abstract class AbstractFloatPlane extends AbstractPlane implements FloatP
     @Override public JsonElement toJson() {
         JsonObject o = new JsonObject();
         o.add("ca", creator().toJson());
-        o.addProperty("type", "float");
+        o.addProperty("type", "double");
         ByteArrayOutputStream bos = new ByteArrayOutputStream();
         try(DataOutputStream dos = new DataOutputStream(bos)) {
             if(getDepth()==0) {
                 for(int i=0;i<getWidth();i++) {
                     for(int j=0;j<getHeight();j++) {
-                        dos.writeFloat(getCell(i,j));
+                        dos.writeDouble(getCell(i,j));
                     }
                 }
             }
@@ -26,7 +26,7 @@ public abstract class AbstractFloatPlane extends AbstractPlane implements FloatP
                 for(int i=0;i<getWidth();i++) {
                     for(int j=0;j<getHeight();j++) {
                         for(int k=0;k<getDepth();k++) {
-                            dos.writeFloat(getCell(i,j,k));
+                            dos.writeDouble(getCell(i,j,k));
                         }
                     }
                 }
@@ -43,8 +43,8 @@ public abstract class AbstractFloatPlane extends AbstractPlane implements FloatP
     @Override public Pen pen() {
         return new Pen() {
             private final int[] _rgb = new int[4];
-            private final float _pal = creator().getPalette().getColorCount()-1f;
-            private final float _mul = (creator().archetype().colors() - 1)/255f;
+            private final double _pal = creator().getPalette().getColorCount()-1f;
+            private final double _mul = (creator().archetype().colors() - 1)/255f;
 
 
             @Override public void setCell(int x, int y, int v) {
@@ -55,7 +55,7 @@ public abstract class AbstractFloatPlane extends AbstractPlane implements FloatP
                 AbstractFloatPlane.this.setCell(x, y, z, v/_pal);
             }
 
-            @Override public void setCell(int x, int y, int z, float v) {
+            @Override public void setCell(int x, int y, int z, double v) {
                 AbstractFloatPlane.this.setCell(x, y, z, v);
             }
 
@@ -84,14 +84,14 @@ public abstract class AbstractFloatPlane extends AbstractPlane implements FloatP
     @Override public Probe probe() {
         if(_probe==null) {
             _probe = new Probe() {
-                final float max = creator().archetype().colors();
+                final double max = creator().archetype().colors();
 
-                @Override public float probe(int x, int y, int z) {
+                @Override public double probe(int x, int y, int z) {
                     return AbstractFloatPlane.this.getCell(x,y,z);
                 }
 
-                @Override public float probeNorm(int x, int y, int z) {
-                    float val = AbstractFloatPlane.this.getCell(x,y,z);
+                @Override public double probeNorm(int x, int y, int z) {
+                    double val = AbstractFloatPlane.this.getCell(x,y,z);
                     return val/max;
                 }
             };
