+++ src/main/org/excelsi/nausicaa/ca/FloatPlane.java
@@ -2,17 +2,17 @@ package org.excelsi.nausicaa.ca;
 
 
 public interface FloatPlane extends Plane {
-    void setCell(int x, int y, float v);
-    void setCell(int x, int y, int z, float v);
+    void setCell(int x, int y, double v);
+    void setCell(int x, int y, int z, double v);
     void setRGBCell(int x, int y, int rgb);
-    float getCell(int x, int y);
-    float getCell(int x, int y, int z);
-    float[] getRow(float[] into, int y, int offset);
-    float[] getBlock(float[] into, int x, int y, int dx, int dy, int offset);
-    float[] getBlock(float[] into, int x, int y, int z, int dx, int dy, int dz, int offset);
-    float[] getCoords(float[] into, int x, int y, int[][] coords, int offset);
-    float[] getCardinal(float[] into, int x, int y, int dx, int dy, int offset);
-    float[] getCardinal(float[] into, int x, int y, int z, int dx, int dy, int dz, int offset);
-    void setRow(float[] row, int y);
-    int computeRgbColor(float idx);
+    double getCell(int x, int y);
+    double getCell(int x, int y, int z);
+    double[] getRow(double[] into, int y, int offset);
+    double[] getBlock(double[] into, int x, int y, int dx, int dy, int offset);
+    double[] getBlock(double[] into, int x, int y, int z, int dx, int dy, int dz, int offset);
+    double[] getCoords(double[] into, int x, int y, int[][] coords, int offset);
+    double[] getCardinal(double[] into, int x, int y, int dx, int dy, int offset);
+    double[] getCardinal(double[] into, int x, int y, int z, int dx, int dy, int dz, int offset);
+    void setRow(double[] row, int y);
+    int computeRgbColor(double idx);
 }
