+++ src/main/org/excelsi/nausicaa/ca/Pattern.java
@@ -10,7 +10,7 @@ public interface Pattern {
     default int next(int pattern, int[] p2, Ctx ctx) {
         throw new UnsupportedOperationException();
     }
-    default float next(int pattern, float[] p2, Ctx ctx) {
+    default double next(int pattern, double[] p2, Ctx ctx) {
         throw new UnsupportedOperationException();
     }
     default Pattern copy() {
