+++ src/main/org/excelsi/nausicaa/ca/Probe.java
@@ -2,6 +2,6 @@ package org.excelsi.nausicaa.ca;
 
 
 public interface Probe {
-    float probe(int x, int y, int z);
-    float probeNorm(int x, int y, int z);
+    double probe(int x, int y, int z);
+    double probeNorm(int x, int y, int z);
 }
